Rename test to testsuite
