+++ pkg/testsuite/protocol_test.go
@@ -1,4 +1,4 @@
-package test
+package testsuite
 
 import (
 	"fmt"
